Splitted code into modules
