+++ index.html
@@ -5,10 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Log in | Hikes Inspire</title>
     <meta name="description" content="Log in page for Hikes Inspire." />
-    <script defer type="module" src="./js/api/constants.mjs"></script>
-    <script defer type="module" src="./js/handlers/auth-handler.mjs"></script>
+    
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-
+    <script defer type="module" src="./js/index.mjs"></script>
+    
+    <link rel="stylesheet" href="/css/theme.css" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -19,8 +20,8 @@
       src="https://kit.fontawesome.com/81a562339c.js"
       crossorigin="anonymous"
     ></script>
-    <link rel="stylesheet" href="/css/theme.css" />
-    <script defer type="module" src="./js/index.mjs"></script>
+  
+   
    
   </head>
   <body>
@@ -142,7 +143,6 @@
                             class="btn btn-outline-warning mt-4 py-2 px-4"
                             id="register-button">New user
                         </button></a>
-                      
                       </div>
                     </div>
             </form>
