--- js/handlers/login.mjs
@@ -1,58 +1,0 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-/*  import { login } from "../api/auth/login.mjs";
-login();
-
-
-export function setLoginFormListener() {
-    const loginForm = document.querySelector("#login-form");
-
-    if (form) {
-
-loginForm.addEventListener("submit", (event) => {
-    event.preventDefault()
-
-    const form = event.target;
-    const formData = new FormData(form);
-
-    const profile = Object.fromEntries(formData.entries())
-
-    //Send it to the API
-    login(profile)
-})
-}
-}
- */
