--- js/handlers/register.mjs
@@ -1,58 +1,0 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-/* import { register } from "../api/auth/register.mjs";
-register();
-
-
-export function setRegisterFormListener() {
-    const registerForm = document.querySelector("#register-form");
-
-    if (form) {
-
-registerForm.addEventListener("submit", (event) => {
-    event.preventDefault()
-
-    const form = event.target;
-    const formData = new FormData(form);
-
-    const profile = Object.fromEntries(formData.entries())
-
-    //Send it to the API
-    register(profile)
-})
-}
-}
- 
- */
