+++ js/routes/home.mjs
@@ -1,0 +1,11 @@
+import { getPosts } from "../api/posts/get.mjs";
+import { setAuthListener } from "../ui/listeners/auth.mjs";
+
+
+
+//Home page function
+export async function homepage() {
+getPosts();
+setAuthListener(); 
+
+}
