+++ js/storage/localstorage.mjs
@@ -1,0 +1,10 @@
+
+//Local Storage
+export function save(key, value) {
+    localStorage.setItem(key, JSON.stringify(value));
+}
+
+export function load(key) {
+    return JSON.parse(localStorage.getItem(key));
+}
+
