+++ js/ui/listeners/auth.mjs
@@ -1,0 +1,8 @@
+
+import { onAuth } from "../events/onAuth.mjs";
+
+//Auth Listener
+export function setAuthListener() {
+    document.forms.auth.addEventListener("submit", onAuth);
+}
+
